adds a build task and updates readme
