+++ config/config.exs
@@ -1,5 +1,5 @@
 use Mix.Config
 
-config :mavlink, dialect: APM
+config :mavlink, dialect: MAVLink.Dialect.APM.Types
 config :mavlink, connections: ["udp:192.168.0.23:14550"]
 config :logger, level: :info
